requirements.txt
pyqrcode
pypng
gtts
